requirements.txt
flask
flask-cors
google-generativeai
gunicorn # Serveur web pour la production